% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_least_squares.R
\name{generalized_least_squares}
\alias{generalized_least_squares}
\title{Generalized least squares}
\usage{
generalized_least_squares(V, X)
}
\arguments{
\item{V}{residual among-species covariance matrix}

\item{X}{model matrix}
}
\value{
the log likelihood value
}
\description{
Generalized least squares
}
\examples{

V <- crossprod(matrix(rnorm(16),4,4))
X <- matrix(1,4,1)

generalized_least_squares(V, X)
}
