% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BM_fitzjohn.R
\name{logl_BM_fitzjohn}
\alias{logl_BM_fitzjohn}
\title{Title}
\usage{
logl_BM_fitzjohn(td, sigma2, mu, trait_name)
}
\arguments{
\item{trait_name}{}
}
\description{
Title
}
\examples{

data("artiodactyla")

logl_BM_fitzjohn(artiodactyla, 0.1, 5.0, "brain_mass_g_log_mean")
}
