% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OU_fitzjohn.R
\name{logl_OU_fitzjohn}
\alias{logl_OU_fitzjohn}
\title{Title}
\usage{
logl_OU_fitzjohn(td, alpha, sigma2, theta, trait_name)
}
\arguments{
\item{trait_name}{}
}
\value{
the log likelihood
}
\description{
Title
}
