% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{loglikelihood}
\alias{loglikelihood}
\title{Calculate the likelihood}
\usage{
loglikelihood(phy, chr.values, par)
}
\arguments{
\item{phy}{phylogeny}

\item{chr.values}{continuous trait values}

\item{par}{parameter vectors}
}
\value{
log-likelihood
}
\description{
Calculate the likelihood
}
