% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_descendants}
\alias{make_descendants}
\title{Title}
\usage{
make_descendants(phy)
}
\arguments{
\item{an}{object of class "phylo"}
}
\value{
a list
}
\description{
Title
}
\examples{

data("artiodactyla")
descendants <- make_descendants(artiodactyla@phylo)
}
