% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OU_fitzjohn.R
\name{logl_OU_fitzjohn}
\alias{logl_OU_fitzjohn}
\title{Calculate the log likelihood of an univariate OU model using the FitzJohn algorithm}
\usage{
logl_OU_fitzjohn(td, alpha, sigma2, theta, trait_name)
}
\arguments{
\item{td}{an object of class "treedata", from the tidytree package. This contains both the tree and the continuous traits}

\item{alpha}{the parameter controlling the rate of adaptation}

\item{sigma2}{the instantaneous diffusion variance parameter}

\item{theta}{the primary optimum}

\item{trait_name}{a string indicating which column in \code{td} that should be used for the trait}
}
\value{
the log likelihood
}
\description{
Calculate the log likelihood of an univariate OU model using the FitzJohn algorithm
}
\examples{
data("artiodactyla")
logl_OU_fitzjohn(artiodactyla, 0.5, 0.5, 5.04, "brain_mass_g_log_mean")
}
