% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OU_vcv.R
\name{logl_OU_vcv}
\alias{logl_OU_vcv}
\title{Title}
\usage{
logl_OU_vcv(td, alpha, sigma2, theta, trait_name)
}
\arguments{
\item{trait_name}{}
}
\description{
Title
}
\examples{

data("artiodactyla")

logl_OU_vcv(artiodactyla, 0.5, 0.5, 5.04, "brain_mass_g_log_mean")
}
