% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phirefly.R
\docType{package}
\name{Phirefly-package}
\alias{Phirefly}
\alias{Phirefly-package}
\title{Phirefly}
\description{
Implements likelihood functions for univariate Brownian motion and Ornstein Uhlenbeck models in comparative phylogenetic methods
}
\author{
\strong{Maintainer}: Haoqing Du \email{duhaoqing115@gmail.com}

Authors:
\itemize{
  \item Priscilla Lau
  \item Bjorn Tore Kopperud
  \item Sebastian Hoehna
}

}
